+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.0.3
 * Add `storage.sync` method
+* Options for call callbacks in reverse insert order (LIFO) [#2](https://github.com/sham-ui/sham-ui-data-storage/issues/2)
 
 ## 0.0.2 
 * Add `browserlist`